requirements copy.txt
reflex==0.2.0
python-dotenv
